:bug: display the final image to debug

+++ submit.py
@@ -63,7 +63,7 @@ def submit(net, writer):
                 plt.title("Predicted")
                 plt.grid(False)
                 plt.subplot(224)
-                plt.imshow(Image.fromarray(np.uint8(masks_pred_np)))
+                plt.imshow(Image.fromarray(masks_pred_np))
                 plt.title("Encoded")
                 plt.grid(False)
                 writer.add_figure(config.PREDICTION_TAG + "/" + str(img_name), F, global_step=index)
